+++ tests/integration/utils.py
@@ -13,14 +13,14 @@ def make_tuple(input: Union[int, tuple[int, ...]]) -> tuple[int, ...]:
     return input
 
 
-def name_func(testcase_func: FunctionType, param_num: int, param: dict) -> str:
+def name_func(testcase_func: FunctionType, _: int, param: dict) -> str:
     """
     Custom name function
 
     Stores in _testMethodName
     """
     if isinstance(param.args[0], FunctionType):
-        return f"{testcase_func.__name__}::{param.args[0].__name__}::{param_num}"
+        return f"{testcase_func.__name__}::{param.args[0].__name__}"
     else:
         # If given list of funcs, use first for naming
-        return f"{testcase_func.__name__}::{param.args[0][0].__name__}::{param_num}"
+        return f"{testcase_func.__name__}::{param.args[0][0].__name__}"
